+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-09 15:22:16
-// Commit: fix(test): prevent double tap duplicate action in test
+// Updated at: 2026-05-09 18:27:05
+// Commit: feat(test): implement responsive test view component
 
 class testComponentState {
-  final String commitRef = '972';
-  final String lastMessage = 'fix(test): prevent double tap duplicate action in test';
-  final DateTime timestamp = DateTime.parse('2026-05-09T15:22:16+05:00');
+  final String commitRef = '974';
+  final String lastMessage = 'feat(test): implement responsive test view component';
+  final DateTime timestamp = DateTime.parse('2026-05-09T18:27:05+05:00');
 }